refactor(ui): standardize component directory structure and import paths
- Update component directory names to use kebab-case
- Update component file names to use PascalCase
- Fix import paths to use @/ alias consistently
- Add path alias configuration for Storybook
- Fix icons exports and imports
