+++ packages/ui/.storybook/main.ts
@@ -1,5 +1,7 @@
 import type { StorybookConfig } from "@storybook/react-vite"
 
+import { resolve } from 'path'
+
 const config: StorybookConfig = {
   stories: ["../src/**/*.mdx", "../src/**/*.stories.@(js|jsx|mjs|ts|tsx)"],
 
@@ -23,6 +25,16 @@ const config: StorybookConfig = {
 
   typescript: {
     reactDocgen: 'react-docgen',
+  },
+
+  viteFinal: async (config) => {
+    if (config.resolve) {
+      config.resolve.alias = {
+        ...config.resolve.alias,
+        '@': resolve(__dirname, '../src'),
+      }
+    }
+    return config
   }
 }
