+++ packages/ui/src/components/ui/calendar/Calendar.tsx
@@ -1,9 +1,9 @@
 import * as React from "react"
-import { Icons } from "./icons"
+import { Icons } from "../icons/Icons"
 import { DayPicker } from "react-day-picker"
 
-import { cn } from "../../lib/utils"
-import { buttonVariants } from "./button"
+import { cn } from "@/lib/utils"
+import { buttonVariants } from "../button/Button"
 
 export type CalendarProps = React.ComponentProps<typeof DayPicker>
