+++ packages/ui/src/components/ui/popover/Popover.tsx
@@ -1,7 +1,7 @@
 import * as React from "react"
 import * as PopoverPrimitive from "@radix-ui/react-popover"
 
-import { cn } from "../../lib/utils"
+import { cn } from "@/lib/utils"
 
 const Popover = PopoverPrimitive.Root
