+++ packages/web/src/components/date-navigation/DateNavigation.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from "@storybook/react"
-import { DateNavigation } from "./date-navigation"
+import { DateNavigation } from "./DateNavigation"
 
 const meta = {
   title: "组件/DateNavigation",
