+++ packages/web/src/components/date-navigation/DateNavigation.tsx
@@ -4,14 +4,14 @@ import { CalendarIcon, ChevronLeftIcon, ChevronRightIcon } from "@radix-ui/react
 import { zhCN } from 'date-fns/locale'
 
 import { cn } from "@/lib/utils"
-import { Button } from "@/components/ui/button"
-import { Calendar } from "@/components/ui/calendar"
+import { Button } from "@/components/ui/button/Button"
+import { Calendar } from "@/components/ui/calendar/Calendar"
 import {
   Popover,
   PopoverContent,
   PopoverTrigger,
-} from "@/components/ui/popover"
-import { buttonVariants } from "@/components/ui/button"
+} from "@/components/ui/popover/Popover"
+import { buttonVariants } from "@/components/ui/button/Button"
 
 interface DateNavigationProps {
   selectedDate: Date
